src/main/java/com/example/songr/controllers/HelloController.java
package com.example.songr.controllers;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.ResponseBody;

@Controller
public class HelloController {

    @GetMapping("/hello")
    @ResponseBody
    public String helloWorld() {
        return "Hello World";
    }

    @GetMapping("/capitalize/{param}")
    @ResponseBody
    public String toStringCaps(@PathVariable String param) {
        return param.toUpperCase();
    }
}
